+++ beetsplug/metasync/__init__.py
@@ -14,28 +14,64 @@
 
 """Synchronize information from music player libraries
 """
+from abc import abstractmethod, ABCMeta
+from importlib import import_module
 
-from beets import ui, logging
+from beets.util.confit import ConfigValueError
+from beets import ui
 from beets.plugins import BeetsPlugin
-from beets.dbcore import types
-from beets.library import DateType
-from sys import modules
-import inspect
 
-# Loggers.
-log = logging.getLogger('beets.metasync')
+
+METASYNC_MODULE = 'beetsplug.metasync'
+
+# Dictionary to map the MODULE and the CLASS NAME of meta sources
+SOURCES = {
+    'amarok': 'Amarok',
+    'itunes': 'Itunes',
+}
+
+
+class MetaSource(object):
+    __metaclass__ = ABCMeta
+
+    def __init__(self, config, log):
+        self.item_types = {}
+        self.config = config
+        self._log = log
+
+    @abstractmethod
+    def sync_from_source(self, item):
+        pass
+
+
+def load_meta_sources():
+    """ Returns a dictionary of all the MetaSources
+    E.g., {'itunes': Itunes} with isinstance(Itunes, MetaSource) true
+    """
+    meta_sources = {}
+
+    for module_path, class_name in SOURCES.items():
+        module = import_module(METASYNC_MODULE + '.' + module_path)
+        meta_sources[class_name.lower()] = getattr(module, class_name)
+
+    return meta_sources
+
+
+META_SOURCES = load_meta_sources()
+
+
+def load_item_types():
+    """ Returns a dictionary containing the item_types of all the MetaSources
+    """
+    item_types = {}
+    for meta_source in META_SOURCES.values():
+        item_types.update(meta_source.item_types)
+    return item_types
 
 
 class MetaSyncPlugin(BeetsPlugin):
 
-    item_types = {
-        'amarok_rating':      types.INTEGER,
-        'amarok_score':       types.FLOAT,
-        'amarok_uid':         types.STRING,
-        'amarok_playcount':   types.INTEGER,
-        'amarok_firstplayed': DateType(),
-        'amarok_lastplayed':  DateType()
-    }
+    item_types = load_item_types()
 
     def __init__(self):
         super(MetaSyncPlugin, self).__init__()
@@ -45,9 +81,9 @@ class MetaSyncPlugin(BeetsPlugin):
                             help='update metadata from music player libraries')
         cmd.parser.add_option('-p', '--pretend', action='store_true',
                               help='show all changes but do nothing')
-        cmd.parser.add_option('-s', '--source', action='store_false',
-                              default=self.config['source'].as_str_seq(),
-                              help="select specific sources to import from")
+        cmd.parser.add_option('-s', '--source', default=[],
+                              action='append', dest='sources',
+                              help='comma-separated list of sources to sync')
         cmd.parser.add_format_option()
         cmd.func = self.func
         return [cmd]
@@ -56,31 +92,43 @@ class MetaSyncPlugin(BeetsPlugin):
         """Command handler for the metasync function.
         """
         pretend = opts.pretend
-        source = opts.source
         query = ui.decargs(args)
 
-        sources = {}
+        sources = []
+        for source in opts.sources:
+            sources.extend(source.split(','))
 
-        for player in source:
-            __import__('beetsplug.metasync', fromlist=[str(player)])
+        sources = sources or self.config['source'].as_str_seq()
 
-            module = 'beetsplug.metasync.' + player
+        meta_source_instances = {}
+        items = lib.items(query)
 
-            if module not in modules.keys():
-                log.error(u'Unknown metadata source \'' + player + '\'')
-                continue
+        # Avoid needlessly instantiating meta sources (can be expensive)
+        if not items:
+            self._log.info(u'No items found matching query')
+            return
 
-            classes = inspect.getmembers(modules[module], inspect.isclass)
+        # Instantiate the meta sources
+        for player in sources:
+            try:
+                meta_source_instances[player] = \
+                    META_SOURCES[player](self.config, self._log)
+            except KeyError:
+                self._log.error(u'Unknown metadata source \'{0}\''.format(
+                    player))
+            except (ImportError, ConfigValueError) as e:
+                self._log.error(u'Failed to instantiate metadata source '
+                                u'\'{0}\': {1}'.format(player, e))
 
-            for entry in classes:
-                if entry[0].lower() == player:
-                    sources[player] = entry[1]()
-                else:
-                    continue
+        # Avoid needlessly iterating over items
+        if not meta_source_instances:
+            self._log.error(u'No valid metadata sources found')
+            return
 
-        for item in lib.items(query):
-            for player in sources.values():
-                player.get_data(item)
+        # Sync the items with all of the meta sources
+        for item in items:
+            for meta_source in meta_source_instances.values():
+                meta_source.sync_from_source(item)
 
             changed = ui.show_model_changes(item)