+++ beetsplug/metasync/amarok.py
@@ -18,12 +18,33 @@
 from os.path import basename
 from datetime import datetime
 from time import mktime
-from beets.util import displayable_path
 from xml.sax.saxutils import escape
-import dbus
+
+from beets.util import displayable_path
+from beets.dbcore import types
+from beets.library import DateType
+from beetsplug.metasync import MetaSource
 
 
-class Amarok(object):
+def import_dbus():
+    try:
+        return __import__('dbus')
+    except ImportError:
+        return None
+
+dbus = import_dbus()
+
+
+class Amarok(MetaSource):
+
+    item_types = {
+        'amarok_rating':      types.INTEGER,
+        'amarok_score':       types.FLOAT,
+        'amarok_uid':         types.STRING,
+        'amarok_playcount':   types.INTEGER,
+        'amarok_firstplayed': DateType(),
+        'amarok_lastplayed':  DateType(),
+    }
 
     queryXML = u'<query version="1.0"> \
                     <filters> \
@@ -31,11 +52,16 @@ class Amarok(object):
                     </filters> \
                 </query>'
 
-    def __init__(self):
+    def __init__(self, config, log):
+        super(Amarok, self).__init__(config, log)
+
+        if not dbus:
+            raise ImportError('failed to import dbus')
+
         self.collection = \
             dbus.SessionBus().get_object('org.kde.amarok', '/Collection')
 
-    def get_data(self, item):
+    def sync_from_source(self, item):
         path = displayable_path(item.path)
 
         # amarok unfortunately doesn't allow searching for the full path, only